add dropout to lstmPeepwhole

+++ pyspark/dl/nn/layer.py
@@ -462,7 +462,7 @@ class GRU(Model):
     creating: createGRU
     '''
 
-    def __init__(self,  input_size, hidden_size, p=0.0, bigdl_type="double"):
+    def __init__(self,  input_size, hidden_size, p=0.0, bigdl_type="float"):
         super(GRU, self).__init__(None, bigdl_type, input_size, hidden_size, p)
